Fix a typing issue in expected mutual information
Some variables were typed as float when they should be double.

+++ sklearn/metrics/cluster/expected_mutual_info_fast.pyx
@@ -20,14 +20,14 @@ np.import_array()
 def expected_mutual_information(contingency, int n_samples):
     """Calculate the expected mutual information for two labelings."""
     cdef int R, C
-    cdef float N, gln_N, emi, term2, term3, gln
+    cdef double N, gln_N, emi, term2, term3, gln
     cdef np.ndarray[double] gln_a, gln_b, gln_Na, gln_Nb, gln_nij, log_Nnij
     cdef np.ndarray[double] nijs, term1
     cdef np.ndarray[double, ndim=2] log_ab_outer
     cdef np.ndarray[np.int32_t] a, b
     #cdef np.ndarray[int, ndim=2] start, end
     R, C = contingency.shape
-    N = float(n_samples)
+    N = np.double(n_samples)
     a = np.sum(contingency, axis=1).astype(np.int32)
     b = np.sum(contingency, axis=0).astype(np.int32)
     # There are three major terms to the EMI equation, which are multiplied to
